Move `get_date_string` to functions.h
